refactor index.ts into several smaller files
