+++ src/config.ts
@@ -84,7 +84,7 @@ export interface GeneTooltipConfig {
   prefetch: 'smart' | 'all' | 'none';
   prefetchThreshold: number;
   truncateSummary: number;
-  theme: string;
+  theme: 'light' | 'dark' | 'auto' | 'material' | 'translucent'| 'light-border' | undefined;
   display: Partial<TooltipDisplayConfig>;
   ideogram: Partial<IdeogramConfig>;
   tippyOptions: Partial<Props>;
@@ -187,3 +187,18 @@ export const defaultConfig: GeneTooltipConfig = {
     placement: 'right',
   }
 };
+
+export function mergeConfig(userConfig: Partial<GeneTooltipConfig> = {}): GeneTooltipConfig {
+    return {
+        ...defaultConfig,
+        ...userConfig,
+        display: {
+            ...defaultConfig.display,
+            ...userConfig.display,
+            links: { ...defaultConfig.display.links, ...userConfig.display?.links },
+        },
+        ideogram: { ...defaultConfig.ideogram, ...userConfig.ideogram },
+        tippyOptions: { ...defaultConfig.tippyOptions, ...userConfig.tippyOptions },
+        nestedTippyOptions: { ...defaultConfig.nestedTippyOptions, ...userConfig.nestedTippyOptions },
+    };
+}
